Separate login functionality related to authorization to authorize utility
